remove vote token, epoch lib

+++ protocol/contracts/interfaces/templegold/IAuctionBase.sol
@@ -2,10 +2,11 @@ pragma solidity ^0.8.20;
 // SPDX-License-Identifier: AGPL-3.0-or-later
 // Temple (interfaces/templegold/IAuctionBase.sol)
 
+
 interface IAuctionBase {
     event Deposit(address indexed depositor, uint256 epochId, uint256 amount);
     event Claim(address indexed user, uint256 epochId, uint256 bidTokenAmount, uint256 auctionTokenAmount);
-    event AuctionStarted(uint256 epochId, address indexed starter, uint64 startTime, uint64 endTime, uint256 auctionTokenAmount);
+    event AuctionStarted(uint256 epochId, address indexed starter, uint128 startTime, uint128 endTime, uint256 auctionTokenAmount);
     
     error CannotDeposit();
     error CannotClaim(uint256 epochId);
@@ -15,9 +16,9 @@ interface IAuctionBase {
 
     struct EpochInfo {
         /// @notice Start time for epoch
-        uint64 startTime;
+        uint128 startTime;
         /// @notice End time for epoch
-        uint64 endTime;
+        uint128 endTime;
         /// @notice Total amount of bid token deposited
         uint256 totalBidTokenAmount;
         /// @notice Total amount of auction tokens to distribute. Constant value
